feat(net,server): Phase 3.7.5 - uint64 character_id end-to-end via Hello
Wire-breaking: kProtocolVersion 1 -> 2 for UDP gameplay protocol.

ServerProtocol
 - HelloMessage.clientNonce: uint32 -> uint64.
 - EncodeHello/DecodeHello use WriteU64/ReadU64, payload 8 -> 12 bytes.

GameplayUdpClient
 - SendHello(uint16, uint16, uint64) signature widened.

Engine
 - EnterWorld override: drop the (& 0xFFFFFFFFu) truncation, propagate the
   full character_id as int64_t in config.
 - InitGameplayNet: charKey read as uint64 (reinterpret-cast preserves bit 63).

Server (shard)
 - HandleHello(uint64) signature widened.
 - ConnectedClient::{helloNonce, persistenceCharacterKey}: uint32 -> uint64.
 - m_bannedCharacterKeys: unordered_set<uint64>.
 - Ban file parser + writer aligned.
 - CharacterPersistence: PersistedCharacterState.characterKey,
   LoadCharacter, BuildCharacterStateRelativePath all uint64. Path filename
   character_<key>.ini accepts std::to_string(uint64) directly.
 - AuctionHouse: sellerCharacterKey, highBidderCharacterKey,
   buyerCharacterKey all uint64 to avoid truncation when storing
   persistenceCharacterKey into auction rows / settlements.
 - RefundGoldToCharacter, DepositMailboxDelivery,
   FindConnectedClientByCharacterKey signatures uint64.

+++ engine/server/AuctionHouse.cpp
@@ -106,14 +106,14 @@ namespace engine::server
 			AuctionListingRecord r{};
 			r.listingId = id;
 			r.sellerClientId = static_cast<uint32_t>(ini.GetInt(p + "seller_client", 0));
-			r.sellerCharacterKey = static_cast<uint32_t>(ini.GetInt(p + "seller_ck", 0));
+			r.sellerCharacterKey = static_cast<uint64_t>(ini.GetInt(p + "seller_ck", 0));
 			r.itemId = static_cast<uint32_t>(ini.GetInt(p + "item", 0));
 			r.quantity = static_cast<uint32_t>(ini.GetInt(p + "qty", 0));
 			r.startBid = static_cast<uint32_t>(ini.GetInt(p + "start", 0));
 			r.buyoutPrice = static_cast<uint32_t>(ini.GetInt(p + "buyout", 0));
 			r.currentBid = static_cast<uint32_t>(ini.GetInt(p + "cur", 0));
 			r.highBidderClientId = static_cast<uint32_t>(ini.GetInt(p + "high_client", 0));
-			r.highBidderCharacterKey = static_cast<uint32_t>(ini.GetInt(p + "high_ck", 0));
+			r.highBidderCharacterKey = static_cast<uint64_t>(ini.GetInt(p + "high_ck", 0));
 			r.expiresAtTick = static_cast<uint32_t>(ini.GetInt(p + "exp_tick", 0));
 			r.closed = ini.GetInt(p + "closed", 0) != 0;
 			m_listings.push_back(r);
